pyproject.toml
[project]
name = "ask-on-data"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rajatnigam", email = "rajatnigam89@gmail.com" }
]
dependencies = [
    "langgraph>=0.3.11",
    "langchain-openai>=0.3.8",
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-community>=0.3.19",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ask_on_data"]

[tool.rye.scripts]
start = "python src/ask_on_data/main.py"